+++ dandy/llm/map/llm_map.py
@@ -8,38 +8,38 @@ from dandy.llm.conf import llm_configs
 from dandy.llm.processor.llm_processor import BaseLlmProcessor
 from dandy.llm.prompt import Prompt
 from dandy.llm.service.config.options import LlmConfigOptions
-from dandy.map.exceptions import MapCriticalException
-from dandy.map.intel import MapValuesIntel
+from dandy.map.exceptions import MapCriticalException, MapRecoverableException
+from dandy.map.intel import MapValuesIntel, MapValueIntel
 from dandy.map.map import Map
 
 
 class BaseLlmMap(BaseLlmProcessor[MapValuesIntel], ABC):
     config: str = 'DEFAULT'
     config_options: LlmConfigOptions = LlmConfigOptions()
-    instructions_prompt: Prompt = (
-        Prompt()
-        .text("Your job is to select a choice for the provided list of choices.")
-        .text("Make sure you select at least one choice that is the most relevant to the users input.")
-    )
+    instructions_prompt: Prompt = Prompt()
     intel_class = MapValuesIntel
+    map_keys_description: str
     map: Map
 
     def __init_subclass__(cls):
         super().__init_subclass__()
 
+        if cls.map_keys_description is None:
+            raise MapCriticalException(f'{cls.__name__} `map_keys_description` is not set.')
+
         if cls.map is None:
-            raise MapCriticalException(f'{cls.__name__} map is not set.')
+            raise MapCriticalException(f'{cls.__name__} `map` is not set.')
 
     @classmethod
     def process(
             cls,
             prompt: Union[Prompt, str],
-            choice_count: int = 1,
+            max_return_values: int | None = None,
     ) -> MapValuesIntel[Any]:
         return cls.process_map_to_intel(
             cls.map,
             prompt,
-            choice_count
+            max_return_values
         )
 
     @classmethod
@@ -47,11 +47,11 @@ class BaseLlmMap(BaseLlmProcessor[MapValuesIntel], ABC):
             cls,
             map: Map,
             prompt: Union[Prompt, str],
-            choice_count: int = 1
+            max_return_values: int | None = None
     ) -> MapValuesIntel[Any]:
         map_values_intel = MapValuesIntel()
 
-        for map_enum in cls.process_prompt_to_intel(map, prompt, choice_count):
+        for map_enum in cls.process_prompt_to_intel(map, prompt, max_return_values):
             map_value = map.get_selected_value(map_enum.value)
 
             if isinstance(map_value, type):
@@ -59,7 +59,7 @@ class BaseLlmMap(BaseLlmProcessor[MapValuesIntel], ABC):
                     map_values_intel.extend(
                         map_value.process(
                             prompt,
-                            choice_count
+                            max_return_values
                         ).items
                     )
                 else:
@@ -70,7 +70,7 @@ class BaseLlmMap(BaseLlmProcessor[MapValuesIntel], ABC):
                     cls.process_map_to_intel(
                         map_value,
                         prompt,
-                        choice_count
+                        max_return_values
                     ).items
                 )
             else:
@@ -83,31 +83,61 @@ class BaseLlmMap(BaseLlmProcessor[MapValuesIntel], ABC):
             cls,
             map: Map,
             prompt: Union[Prompt, str],
-            choice_count: int = 1,
+            max_return_values: int | None = None,
     ) -> MapValuesIntel:
 
-        system_prompt = (
-            Prompt()
-            .prompt(cls.instructions_prompt)
-            .text(f'Please select {choice_count} of the following choices by number using the following rules.')
-            .line_break()
-            .sub_heading('Rules:')
-            .list([
-                'Select the choice that best matches the users input.',
-                'Return at least one choice by number.'
-            ])
-            .line_break()
-            .sub_heading('Choices:')
-            .text(map.keyed_choices_str())
-        )
+        if max_return_values is not None and max_return_values > 1:
+            key_str = 'keys'
+            intel_class = MapValuesIntel[cls.map.as_enum()]
+        else:
+            key_str = 'key'
+            intel_class = MapValueIntel[cls.map.as_enum()]
 
-        return llm_configs[cls.config].generate_service(
+        system_prompt = Prompt()
+        system_prompt.prompt(cls.instructions_prompt)
+        system_prompt.text(f'You\'re an "{cls.map_keys_description}" assistant')
+
+        system_prompt.line_break()
+
+        system_prompt.text(
+            f'Read through all of the "{cls.map_keys_description}" and return the numbered {key_str} that match information relevant to the user\'s input.')
+
+        system_prompt.line_break()
+
+        if max_return_values is not None and max_return_values > 0:
+            if max_return_values == 1:
+                system_prompt.text(f'You must return exactly one numbered {key_str}.')
+            else:
+                system_prompt.text(f'Return up to a maximum of {max_return_values} numbered {key_str} and return at least one at a minimum.')
+        else:
+            system_prompt.text(f'Return the numbered {key_str} you find that are most relevant and return at least one.')
+
+        system_prompt.line_break()
+        system_prompt.heading(f'"{cls.map_keys_description}"')
+        system_prompt.line_break()
+
+        system_prompt.dict(map.keyed_choices_dict)
+
+        return_values_intel = llm_configs[cls.config].generate_service(
             llm_options=cls.config_options
         ).process_prompt_to_intel(
             prompt=prompt if isinstance(prompt, Prompt) else Prompt(prompt),
-            intel_class=MapValuesIntel[cls.map.as_enum()],
+            intel_class=intel_class,
             system_prompt=system_prompt
         )
+
+        if isinstance(return_values_intel, MapValueIntel):
+            return_values_intel = MapValuesIntel[cls.map.as_enum()](
+                items=[return_values_intel.item.value]
+            )
+
+        if len(return_values_intel) == 0:
+            raise MapRecoverableException(f'No {cls.map_keys_description} found.')
+
+        if max_return_values is not None and len(return_values_intel) > max_return_values:
+            raise MapRecoverableException(f'Too many {cls.map_keys_description} found.')
+
+        return return_values_intel
 
     @classmethod
     def process_to_future(cls, *args, **kwargs) -> AsyncFuture[MapValuesIntel]: