+++ docs/changelog/changelog.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## v0.18.0
+
+### Features
+
+- The `BaseLlmMap` class now has a new required attribute `map_keys_description` that drastically improves accuracy.
+
+### Changes
+
+- On the `BaseLlmMap.process` the argument `choice_count` was renamed to `max_return_values` to make it more clear what it does and is now optional.
+- The `BaseListIntel` now requires you to set the type of the `items` class variable with a type hint.
+
+### Fixes
+
+- Changed `Prompt` into a dataclass to allow typing with `BaseIntel`, `BaseListIntel` and `BaseModel`.
+- Fixed a problem with the way `BaseLlmMap` was prompting for single responses using `max_return_values=1`.
+
 ## v0.17.0
 
 ### Features